update code to latest templates
