+++ samples/csharp/getting-started/WeatherMicroservice/Extensions.cs
@@ -1,8 +1,3 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
 namespace WeatherMicroservice
 {
     public static class Extensions
@@ -11,9 +6,13 @@ namespace WeatherMicroservice
         {
             double result;
             if (double.TryParse(input, out result))
+            {
                 return result;
+            }
             else
+            {
                 return default(double?);
+            }
         }
     }
 }
