+++ samples/csharp/getting-started/WeatherMicroservice/README.md
@@ -1,27 +1,39 @@
-C# Microservices Sample
-================
+# Welcome to ASP.NET Core
 
-This sample is created during the [Creating Microservices hosted in Docker](https://docs.microsoft.com/dotnet/articles/csharp/tutorials/microservices)
-tutorial. Please see that topic for detailed steps on the code
-for this sample.
+We've made some big updates in this release, so it’s **important** that you spend a few minutes to learn what’s new.
 
-Key Features
-------------
+You've created a new ASP.NET Core project. [Learn what's new](https://go.microsoft.com/fwlink/?LinkId=518016)
 
-This sample demonstrates creating a web service using asp.net core and C#
-and deploying that service in a Docker Container.
+## This application consists of:
 
-Build and Run
--------------
+*   Sample pages using ASP.NET Core MVC
+*   [Bower](https://go.microsoft.com/fwlink/?LinkId=518004) for managing client-side libraries
+*   Theming using [Bootstrap](https://go.microsoft.com/fwlink/?LinkID=398939)
 
-To build and run the sample locally, type the following three commands:
+## How to
 
-`dotnet restore`
-`dotnet build`
-`dotnet run`
+*   [Add a Controller and View](https://go.microsoft.com/fwlink/?LinkID=398600)
+*   [Add an appsetting in config and access it in app.](https://go.microsoft.com/fwlink/?LinkID=699562)
+*   [Manage User Secrets using Secret Manager.](https://go.microsoft.com/fwlink/?LinkId=699315)
+*   [Use logging to log a message.](https://go.microsoft.com/fwlink/?LinkId=699316)
+*   [Add packages using NuGet.](https://go.microsoft.com/fwlink/?LinkId=699317)
+*   [Add client packages using Bower.](https://go.microsoft.com/fwlink/?LinkId=699318)
+*   [Target development, staging or production environment.](https://go.microsoft.com/fwlink/?LinkId=699319)
 
-`dotnet restore` installs all the dependencies for this sample into the current directory.
-`dotnet build` creates the output assembly (or assemblies).
-`dotnet run` runs the output assembly.
+## Overview
 
-To install and run the sample in Docker, consult the [full tutorial](../../../docs/csharp/tutorials/microservices.md).
+*   [Conceptual overview of what is ASP.NET Core](https://go.microsoft.com/fwlink/?LinkId=518008)
+*   [Fundamentals of ASP.NET Core such as Startup and middleware.](https://go.microsoft.com/fwlink/?LinkId=699320)
+*   [Working with Data](https://go.microsoft.com/fwlink/?LinkId=398602)
+*   [Security](https://go.microsoft.com/fwlink/?LinkId=398603)
+*   [Client side development](https://go.microsoft.com/fwlink/?LinkID=699321)
+*   [Develop on different platforms](https://go.microsoft.com/fwlink/?LinkID=699322)
+*   [Read more on the documentation site](https://go.microsoft.com/fwlink/?LinkID=699323)
+
+## Run & Deploy
+
+*   [Run your app](https://go.microsoft.com/fwlink/?LinkID=517851)
+*   [Run tools such as EF migrations and more](https://go.microsoft.com/fwlink/?LinkID=517853)
+*   [Publish to Microsoft Azure Web Apps](https://go.microsoft.com/fwlink/?LinkID=398609)
+
+We would love to hear your [feedback](https://go.microsoft.com/fwlink/?LinkId=518015)
