+++ samples/csharp/getting-started/WeatherMicroservice/Startup.cs
@@ -6,8 +6,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using System.IO;
-using System.Text;
+using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 
 namespace WeatherMicroservice
@@ -15,19 +14,26 @@ namespace WeatherMicroservice
     public class Startup
     {
         // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit http://go.microsoft.com/fwlink/?LinkID=398940
+        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app)
+        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
+            loggerFactory.AddConsole();
+
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+
             app.Run(async (context) =>
             {
                 var latString = context.Request.Query["lat"].FirstOrDefault();
                 var longString = context.Request.Query["long"].FirstOrDefault();
-
+                
                 var latitude = latString.TryParse();
                 var longitude = longString.TryParse();
 
@@ -38,31 +44,15 @@ namespace WeatherMicroservice
                     {
                         forecast.Add(new WeatherReport(latitude.Value, longitude.Value, days));
                     }
-
                     var json = JsonConvert.SerializeObject(forecast, Formatting.Indented);
                     context.Response.ContentType = "application/json; charset=utf-8";
                     await context.Response.WriteAsync(json);
                 }
                 else
                 {
-                    await context.Response.WriteAsync($"Retrieving Weather for lat: {latitude}, long: {longitude}");
+                    await context.Response.WriteAsync($"Retrieving Weather for lat: {latitude}, long: {longitude}");                
                 }
             });
         }
-
-        // Entry point for the application.
-
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder() 
-                .UseKestrel() 
-                .UseIISIntegration() 
-                .UseStartup<Startup>() 
-                .UseUrls("http://0.0.0.0:80")
-                .Build(); 
- 
-           host.Run(); 
-            
-        } 
     }
 }
